Move validator `err.rs` to `schema/err.rs`
This only contains schema parsing errors, so it makes sense to move it
out of the crate root
